requirements.txt
llama-cpp-python
faiss-cpu
sentence-transformers
numpy==1.26.4